Added location and social medias support for second follow ups command
