+++ app/Console/Commands/Hlp/Notify/OrganisationAdminInvitee/SecondFollowUpsCommand.php
@@ -6,6 +6,7 @@ use App\Emails\OrganisationAdminInviteSecondFollowUps\NotifyInviteeEmail;
 use App\Generators\AdminUrlGenerator;
 use App\Models\Notification;
 use App\Models\OrganisationAdminInvite;
+use App\Transformers\OrganisationInviteTransformer;
 use Illuminate\Console\Command;
 use Illuminate\Support\Facades\Date;
 use Illuminate\Support\Facades\DB;
@@ -30,14 +31,16 @@ class SecondFollowUpsCommand extends Command
      * Execute the console command.
      *
      * @param \App\Generators\AdminUrlGenerator $adminUrlGenerator
+     * @param \App\Transformers\OrganisationInviteTransformer $transformer
      */
-    public function handle(AdminUrlGenerator $adminUrlGenerator): void
+    public function handle(AdminUrlGenerator $adminUrlGenerator, OrganisationInviteTransformer $transformer): void
     {
         $dates = array_map(function (int $week): string {
             return Date::today()->subWeeks($week)->toDateString();
         }, range(5, 9));
 
         $organisationAdminInvites = OrganisationAdminInvite::query()
+            ->with('organisation', 'organisation.location', 'organisation.socialMedias')
             ->whereNotNull('email')
             ->whereIn(DB::raw('cast(`created_at` as date)'), $dates)
             ->get();
@@ -48,11 +51,15 @@ class SecondFollowUpsCommand extends Command
                     $organisationAdminInvite->email,
                     [
                         'ORGANISATION_NAME' => $organisationAdminInvite->organisation->name,
-                        'ORGANISATION_ADDRESS' => 'N/A', // TODO
+                        'ORGANISATION_ADDRESS' => $transformer->transformAddress(
+                            $organisationAdminInvite->organisation->location
+                        ) ?: 'N/A',
                         'ORGANISATION_URL' => $organisationAdminInvite->organisation->url ?: 'N/A',
                         'ORGANISATION_EMAIL' => $organisationAdminInvite->organisation->email ?: 'N/A',
                         'ORGANISATION_PHONE' => $organisationAdminInvite->organisation->phone ?: 'N/A',
-                        'ORGANISATION_SOCIAL_MEDIA' => 'N/A', // TODO
+                        'ORGANISATION_SOCIAL_MEDIA' => $transformer->transformSocialMedias(
+                            $organisationAdminInvite->organisation->socialMedias
+                        ) ?: 'N/A',
                         'ORGANISATION_DESCRIPTION' => $organisationAdminInvite->organisation->description,
                         'INVITE_URL' => $adminUrlGenerator->generateOrganisationAdminInviteUrl(
                             $organisationAdminInvite
